chore: Add files to ETL setup
Firts commit to deal the ETL process. In the folder, there are
differente file types. The first are CSV files with data to be
processed, the second are some testing code that cover some
cmd data recovering. The last is the file with production code.

The prodution code is responsible for ETL process, loading the
csv's data, tranform it in the desired format and then load them
to the database.